refactor: browserのテストツールをファイル移動
