+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/bullet-list/__tests__/input-rule.browser.spec.ts
@@ -5,7 +5,7 @@ import Text from '@tiptap/extension-text';
 import { userEvent } from '@vitest/browser/context';
 import { describe, expect, it } from 'vitest';
 import { waitSelectionChange } from '../../../../tests/dom';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { BulletList } from '..';
 
 const ListItem = TiptapListItem.extend({
