+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/code-block-container/diff-code-block/__tests__/input-rule.browser.spec.ts
@@ -5,7 +5,7 @@ import Text from '@tiptap/extension-text';
 import { userEvent } from '@vitest/browser/context';
 import { describe, expect, it } from 'vitest';
 import { waitSelectionChange } from '../../../../../tests/dom';
-import { renderTiptapEditor } from '../../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../../tests/browser/editor';
 import { CodeBlock } from '../../code-block';
 import { CodeBlockFileName } from '../../code-block-file-name';
 import { CodeBlockContainer } from '../../index';
