+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/details/__tests__/open.browser.spec.ts
@@ -3,7 +3,7 @@ import Paragraph from '@tiptap/extension-paragraph';
 import Text from '@tiptap/extension-text';
 import { page, userEvent } from '@vitest/browser/context';
 import { describe, expect, it } from 'vitest';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { Details } from '..';
 import { DetailsContent } from '../content';
 import { DetailsSummary } from '../summary';
