+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/figure/__tests__/key.browser.spec.ts
@@ -5,7 +5,7 @@ import { userEvent } from '@vitest/browser/context';
 import { describe, expect, it } from 'vitest';
 import LakeImage from '../../../../tests/assets/sikotuko.jpeg';
 import { waitSelectionChange } from '../../../../tests/dom';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { Caption } from '../caption';
 import { Image } from '../image';
 import { Figure } from '../index';
