+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/heading/__tests__/toc-plugin.spec.ts
@@ -2,7 +2,7 @@ import Document from '@tiptap/extension-document';
 import Paragraph from '@tiptap/extension-paragraph';
 import Text from '@tiptap/extension-text';
 import { describe, expect, it, vi } from 'vitest';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import Heading from '..';
 
 const basicExtension = [Document, Paragraph, Text, Heading];
