+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/message/__tests__/command.spec.ts
@@ -4,7 +4,7 @@ import { ListKit } from '@tiptap/extension-list';
 import Paragraph from '@tiptap/extension-paragraph';
 import Text from '@tiptap/extension-text';
 import { describe, expect, it } from 'vitest';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { CodeBlockContainer } from '../../code-block-container';
 import { CodeBlock } from '../../code-block-container/code-block';
 import { CodeBlockFileName } from '../../code-block-container/code-block-file-name';
