+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/message/__tests__/key.browser.spec.ts
@@ -4,7 +4,7 @@ import Text from '@tiptap/extension-text';
 import { userEvent } from '@vitest/browser/context';
 import { describe, expect, it } from 'vitest';
 import { waitSelectionChange } from '../../../../tests/dom';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { Message } from '../message';
 import { MessageContent } from '../message-content';
