+++ packages/zenn-wysiwyg-editor/src/extensions/nodes/message/__tests__/render.spec.ts
@@ -2,7 +2,7 @@ import Document from '@tiptap/extension-document';
 import Paragraph from '@tiptap/extension-paragraph';
 import Text from '@tiptap/extension-text';
 import { describe, expect, it } from 'vitest';
-import { renderTiptapEditor } from '../../../../tests/editor';
+import { renderTiptapEditor } from '../../../../tests/browser/editor';
 import { Message } from '../message';
 import { MessageContent } from '../message-content';
